Added rush stuff and d05

+++ rush00/src/install.php
@@ -2,19 +2,24 @@
 <?php
 include 'getdata.php';
 
-mkdir("../private");
-unlink(USRFILE);
-file_put_contents(USRFILE, serialize_(array()));
-unlink(BASKETFILE);
-file_put_contents(BASKETFILE, serialize_(array()));
+if (!file_exists("../private"))
+	mkdir("../private");
+if (file_exists(USRFILE))
+	unlink(USRFILE);
+if (file_exists(BASKETFILE))
+	unlink(BASKETFILE);
 
-unlink(ITEMFILE);
+if (file_exists(ITEMFILE))
+	unlink(ITEMFILE);
 $items = array(
-	'apple' => array('name' => 'apple', 'price' => 15, 'tags' => array('fruit', 'apple', 'food'), 'img' => 'https://www.markon.com/sites/default/files/styles/large/public/pi_photos/Apples_Golden_Delicious_Hero.jpg'),
-	'pear' => array('name' => 'pear', 'price' => 12, 'tags' => array('fruit', 'pear', 'food'), 'img' => 'https://cdn1.medicalnewstoday.com/content/images/hero/285/285430/285430_1100.jpg'),
-	'bicycle' => array('name' => 'bicycle', 'price' => 110, 'tags' => array('exercise', 'activity', 'bike'), 'img' => 'http://cdn.shopify.com/s/files/1/1245/1481/products/Hero_600_square2_1024x1024.jpg?v=1546547634'),
-	'passionfruit' => array('name' => 'passionfruit', 'price' => 12, 'tags' => array('fruit', 'passionfruit', 'passion', 'food'), 'img' => 'https://www.rachaelraymag.com/.image/t_share/MTQ1NzM5NjAxNjE1NTI5OTA0/passion-fruit-half-102951784.jpg'),
+	'mcintosh' => array('name' => 'mcintosh', 'price' => 2, 'tags' => array('red', 'blue', 'tangy', 'tart', 'applesauce', 'snack', 'pie'), 'img' => 'https://www.stemilt.com/wp-content/uploads/2016/07/McIntosh.png'),
+	'fuji' => array('name' => 'fuji', 'price' => 3, 'tags' => array('red', 'yellow', 'year', 'popular', 'sweet', 'firm'), 'img' => 'https://images-na.ssl-images-amazon.com/images/I/81pO3uqRDML._SY355_.jpg'),
+	'red_delicious' => array('name' => 'red_delicious', 'price' => 2, 'tags' => array('red', 'delicious'), 'img' => 'http://usapple.org/wp-content/uploads/2016/02/apple-red-delicious-337x335.png'),
+	'gala' => array('name' => 'gala', 'price' => 2.5, 'tags' => array('red', 'orange', 'crisp', 'juicy', 'year'), 'img' => 'http://usapple.org/wp-content/uploads/2016/02/apple-gala-337x335.png'),
+	'crispin' => array('name' => 'crispin', 'price' => 2, 'tags' => array('crisp', 'mutsu', 'green', 'sweet'), 'img' => 'http://ad.thevictorapplefarm.com/images/ourapples/crispinpage.jpg'),
+	'golden_delicious' => array('name' => 'golden_delicious', 'price' => 3, 'tags' => array('golden', 'delicious', 'gold', 'best', 'insane', 'fantastic', 'orgasmic'), 'img' => 'http://usapple.org/wp-content/uploads/2016/02/apple-golden-delicious-337x335.png'),
 );
 file_put_contents(ITEMFILE, serialize_($items));
 
-unlink(BASKETFILE);
+if (file_exists(BASKETFILE))
+	unlink(BASKETFILE);
